Set up condo sales so we can make real money. Added sale price to condos. Updated person to randomly have a higher occupation and afford a condo. Sped up elevators to keep time. Adjusted camera controls to prevent bounce.
git-svn-id: https://svn.code.sf.net/p/highrisedev/code@182 f146f8d7-247d-460d-a1b3-d0a612191574

+++ Window/Interface.cpp
@@ -120,7 +120,7 @@ void Interface::Update (float dt)
    mClock.Update(1); // 1 minute update
    if( count < 1)
    {
-      count = dt;
+      count = (int)dt;
       if( mClock.GetDayOfYear() != mCurDay )
       {
          mStats.SetDayOfWeek (mClock.DayOfWeekToString());
